% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmBART-package.R
\docType{package}
\name{pmBART-package}
\alias{pmBART}
\alias{pmBART-package}
\title{pmBART: probit monotone BART}
\description{
\code{pmBART} implements probit monotone BART for binary outcomes. All predictors
have non-decreasing effects.
}
\author{
\strong{Maintainer}: Jared Fisher \email{fisher@stat.byu.edu}

Authors:
\itemize{
  \item Jared Fisher \email{fisher@stat.byu.edu}
}

}
\keyword{internal}
